% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ar1Cpp}
\alias{ar1Cpp}
\title{AR1 correlation matrix function}
\usage{
ar1Cpp(t, alpha)
}
\arguments{
\item{t}{Specify matrix dimensions}

\item{alpha}{Specify the correlation parameter}
}
\value{
AR1 correlation matrix with t_dimensions and alpha_parameter
}
\description{
generate autoregression working correlation matrix
}
\examples{
\dontrun{
t <- 5
alpha <- 0.5
ar1Cpp(t,alpha)
}
}
