% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{exchCpp}
\alias{exchCpp}
\title{EX correlation matrix function}
\usage{
exchCpp(t, alpha)
}
\arguments{
\item{t}{Specify matrix dimensions}

\item{alpha}{Specify the correlation parameter}
}
\value{
EX correlation matrix with t_dimensions and alpha_parameter
}
\description{
generate exchangable working correlation matrix
}
\examples{
\dontrun{
t <- 5
alpha <- 0.5
exchCpp(t,alpha)
}
}
