% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SA24204141_R.R
\name{GEE_EL}
\alias{GEE_EL}
\title{GEE_EL function}
\usage{
GEE_EL(y, x, cor = "EC", init.beta = NULL, maxit = 1000)
}
\arguments{
\item{y}{Response Variables (Matrix)}

\item{x}{Independent variables (3-dimensional arrays)}

\item{cor}{Designated the type of working correlation;Optional types are:ID、EC、AR、ST}

\item{init.beta}{initial value of iteration (vector)}

\item{maxit}{Maximum number of iteration (int)}
}
\value{
final coefficient estimates
}
\description{
Estimates of regression coefficients are obtained by maximizing the empirical likelihood ratio
}
\examples{
\dontrun{
#data generation
CC <- matrix(c(1, 0.4, 0.4, 1), 2, 2)
zz <- mvrnorm(1000, c(0, 0), CC)
x1 <- matrix(zz[,1], 100, 10)
x2 <- matrix(zz[,2], 100, 10)
X0 <- array(NA, c(100, 10, 3))
X0[,,1] <- 1
X0[,,2] <- x1
X0[,,3] <- x2
Beta <- c(1,-2,1)
tMu <- matrix(NA, 100, 10)
for(i in 1:100){
  tMu[i,] <- as.vector(X0[i,,]\%*\%Beta)
}
Prob <- logit(tMu)
tR <- matrix(0.5, 10, 10)
diag(tR) <- 1
Y0 <- matrix(NA, 100, 10)
for(i in 1:100){
  Y0[i,] <- as.vector(rmvbin(1, margprob=Prob[i,], sigma=tR))
}
#use GEE function
GEE_EL(Y0,X0,cor = "EC")
}
}
